Move GLProgram to its own module.
