+++ gl/Main.hs
@@ -18,6 +18,7 @@ import Foreign.Storable
 import Graphics.GL.Core41
 import Graphics.Shader.Fragment
 import GL.Exception
+import GL.Program
 import GL.Shader.Fragment
 import Linear.V3
 import Prelude hiding (IO)
